+++ gemstone/auth/validation_strategies/__init__.py
@@ -1,2 +1,0 @@
-from .basic_cookie_strategy import BasicCookieStrategy
-from .header_strategy import HeaderValidationStrategy