+++ tests/services/service_jsonrpc_specs.py
@@ -1,4 +1,4 @@
-from gemstone import MicroService, public_method, private_api_method
+from gemstone import MicroService, exposed_method
 
 
 class ServiceJsonRpcSpecs(MicroService):
@@ -9,30 +9,27 @@ class ServiceJsonRpcSpecs(MicroService):
 
     skip_configuration = True
 
-    @public_method
+    @exposed_method()
     def subtract(self, a, b):
         if not isinstance(a, int) or not isinstance(b, int):
             raise TypeError("Arguments must be integers")
         return a - b
 
-    @public_method
+    @exposed_method()
     def sum(self, *args):
         return sum(args)
 
-    @public_method
+    @exposed_method()
     def update(self, a):
         return str(a)
 
-    @public_method
+    @exposed_method()
     def get_data(self):
         return ["hello", 5]
 
-    @public_method
+    @exposed_method()
     def notify_hello(self, a):
         return a
-
-    def api_token_is_valid(self, api_token):
-        return api_token == "test-token"
 
 
 if __name__ == '__main__':